do not link libc++ statically when clang presents
Signed-off-by: Schrodinger ZHU Yifan <i@zhuyi.fan>

As it is stated before, we are using libtiflash-proxy and tiflash together. Both of them
rely on libstdc++/libc++, which are previously statically linked.

This is a dangerous situation since both tiflash and libtiflash-proxy will contain an implementation
of STL so symbols with the same name will duplicate. It may be the case that an exception thrown in
libtiflash-proxy should be handled within libtiflash-proxy's context; however, it is somehow forwarded
to tiflash's function (because they have the same symbol). This situation is the so-called ODR violation
[1]. A good explain for this is available at [2].

It is fortunate that we did not encounter any problem before. Nevertheless, being fortunate does not mean
that we are doing it right. For safety, we are not going to change this in GCC pipeline. However, as we are
to migrate to LLVM, let us prepare dynamic linkage for the future. This PR together with #3683 should be able
to clean up ODR violation for crtical runtime libs. It is hard to judge whether other dependencies like gRPC
are facing the same problem; but let us do what we can do and what we should do for the time being.

This PR will not affect existing pipelines:

- GCC's pipeline will be kept as before.
- AppleClang pipeline is already not using this flag.

[1]: https://en.cppreference.com/w/cpp/language/definition
[2]: https://developer.android.com/ndk/guides/cpp-support#sr

Signed-off-by: Schrodinger ZHU Yifan <i@zhuyi.fan>