[Bug] Fix the empty values when runtime filter is LOCAL mode
Fixed pingcap/tidb#45300

This PR change the IN RS operator when values is empty.
Using return RSResult::None instead of thrown exception when values is empty.

The In RS Operator is only used for runtime filter.
When filter is empty (the build side is empty infact), the IN values will be empty.
So this is not a exception and it should skip all of scan of packs.